+++ src/components/form/field_search/field_search.tsx
@@ -1,4 +1,4 @@
-import React, { Component, InputHTMLAttributes } from 'react';
+import React, { Component, InputHTMLAttributes, KeyboardEvent } from 'react';
 import classNames from 'classnames';
 import { Browser } from '../../../services/browser';
 import { ENTER } from '../../../services/key_codes';
@@ -53,13 +53,24 @@ export interface EuiFieldSearchProps
   append?: EuiFormControlLayoutProps['append'];
 }
 
-export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
+interface EuiFieldSearchState {
+  value: string;
+}
+
+export class EuiFieldSearch extends Component<
+  EuiFieldSearchProps,
+  EuiFieldSearchState
+> {
   static defaultProps = {
     fullWidth: false,
     isLoading: false,
     incremental: false,
     compressed: false,
     isClearable: true,
+  };
+
+  state = {
+    value: this.props.value || '',
   };
 
   inputElement: HTMLInputElement | null = null;
@@ -124,6 +135,7 @@ export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
       // set focus on the search field
       this.inputElement.focus();
     }
+    this.setState({ value: '' });
   };
 
   componentWillUnmount() {
@@ -138,10 +150,11 @@ export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
   };
 
   onKeyUp = (
-    event: React.KeyboardEvent<HTMLInputElement>,
+    event: KeyboardEvent<HTMLInputElement>,
     incremental?: boolean,
     onSearch?: (value: string) => void
   ) => {
+    this.setState({ value: (event.target as HTMLInputElement).value });
     if (this.props.onKeyUp) {
       this.props.onKeyUp(event);
       if (event.defaultPrevented) {
@@ -159,7 +172,6 @@ export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
       id,
       name,
       placeholder,
-      value,
       isInvalid,
       fullWidth,
       isLoading,
@@ -173,6 +185,7 @@ export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
       ...rest
     } = this.props;
 
+    const { value } = this.state;
     const classes = classNames(
       'euiFieldSearch',
       {
@@ -190,7 +203,7 @@ export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
         fullWidth={fullWidth}
         isLoading={isLoading}
         clear={
-          isClearable && value && !rest.readOnly && !rest.disabled
+          isClearable && value.length > 0 && !rest.readOnly && !rest.disabled
             ? { onClick: this.onClear }
             : undefined
         }
@@ -204,7 +217,6 @@ export class EuiFieldSearch extends Component<EuiFieldSearchProps> {
             name={name}
             placeholder={placeholder}
             className={classes}
-            value={value}
             onKeyUp={e => this.onKeyUp(e, incremental, onSearch)}
             ref={this.setRef}
             {...rest}